Remake io, builders, and added matlab functions
Io is further developed to handle output from matlab in -v7.3 format.

Builders are developed to import multem output better. They are not tested properly yet.

Several matlab functions are added to set up MULTEM simulations. Example multem scripts are also provided, along with a test-model with broken symmetry. In the future, a test model with a monoclinic unit cell should also be made to test symmetries in CBED and diffraction data.